Update auto SPI for timestamp changes
The 2019 FPGA image switched the output of auto SPI from plain bytes to a
sequence of 32-bit words (timestamp, then words with the byte values in the
least significant byte of each word).

In addition to changing the HAL and simulators to reflect this, add piecewise
integration support to wpilibc/wpilibj SPI to take advantage of the timestamps
and use it in the ADXRS450 gyro.

+++ hal/src/main/java/edu/wpi/first/hal/sim/SpiReadAutoReceiveBufferCallback.java
@@ -8,5 +8,5 @@
 package edu.wpi.first.hal.sim;
 
 public interface SpiReadAutoReceiveBufferCallback {
-  int callback(String name, byte[] buffer, int numToRead);
+  int callback(String name, int[] buffer, int numToRead);
 }
